Split production-readiness validation modules
